Fix r10k to be more dynamic for hosts
Set the configure_r10k.pp and Puppetfile configurations to not specify a
cloning user and expect /root/.ssh/config to have this appropriately
setup. This will allow us to share these files across environments in a
cleaner manner, otherwise we'll have to share an account for pulls
because of how the Puppetfile is used.

Change-Id: I7dfec39bf12673874ae2a27a7420b47aeac874bc
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>