Add in vim and timezone modules
Add in ghoneycutt-vim and saz-timezone modules

This will allow us to manage installing and configuring vim in a better
manner.

This will allow us to easily manage / force the system TZ. The default
is to set to UTC if nothing is defined in hiera which is exactly what we
want. This gets around the fact that Rackspace and our other cloud
providers may default to a different TZ.

Change-Id: I8b2baf1bd518e68d3972d81d98f29c4da8212f56
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>