Add dhoppe/screen module
We don't currently force an install of screen. This is the only module
on the Forge and thankfully it actually looks decent. They state they
haven't tested on anything but Ubuntu and Debian but my quick look over
of the module looks like it should work on EL.

Temporarily utilize a forked GitHub repo as upstream does not presently
support installation on RedHat family (explicit fail). Current tests
indicate that everything will work fine if RedHat is allowed to install,
no extra changes needed.

Change-Id: I5f6b769c723191780d05a7f727205b28dc8a0e35
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>