Update firewall module to v1.7.1
Puppetlabs-firewall was updated to v1.7.1 back in August. This version
gets rid of the failed first-time runs we see coming out of the
profile::bacula::client (I don't think it was that profile actually
causing the problem, just the one exposing it). v1.7.1 also deprecates
the use of the ambiguous port option for firewall rules. As such, all of
the profiles that do firewall rules will need to be updated to use
dport/sport as appropriate to get rid of the warnings.

Change-Id: Ife5fdc6ed0309735f35aa17d4220719c9538f91e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>